requirements.txt
# Core Dependencies
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2
xlrd==2.0.1

# Data Processing & Analysis
scikit-learn==1.3.2
scipy==1.11.4

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Web Framework
streamlit==1.29.0
streamlit-aggrid==0.3.4.post3

# AI/ML
google-generativeai==0.8.3
langchain==0.1.0

# Database
sqlalchemy==2.0.23

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
